Fixed integer tests

+++ knarr/src/test/java/edu/gmu/swe/test/knarr/StackOpITCase.java
@@ -125,17 +125,17 @@ public class StackOpITCase {
 		assertFalse(ret.isEmpty());
 		for (SimpleEntry<String, Object> e : ret) {
 			switch (e.getKey()) {
-				case "d1":  assertEquals((long)res1 , e.getValue());  break;
-				case "d2":  assertEquals((long)res2 , e.getValue());  break;
-				case "d3":  assertEquals((long)res3 , e.getValue());  break;
-				case "d4":  assertEquals((long)res4 , e.getValue());  break;
-				case "d5":  assertEquals((long)res5 , e.getValue());  break;
-				case "d6":  assertEquals((long)res6 , e.getValue());  break;
-				case "d7":  assertEquals((long)res7 , e.getValue());  break;
-				case "d8":  assertEquals((long)res8 , e.getValue());  break;
-				case "d9":  assertEquals((long)res9 , e.getValue());  break;
-				case "d10": assertEquals((long)res10, e.getValue()); break;
-				case "d11": assertEquals((long)res11, e.getValue()); break;
+				case "d1":  assertEquals(res1 , e.getValue());  break;
+				case "d2":  assertEquals(res2 , e.getValue());  break;
+				case "d3":  assertEquals(res3 , e.getValue());  break;
+				case "d4":  assertEquals(res4 , e.getValue());  break;
+				case "d5":  assertEquals(res5 , e.getValue());  break;
+				case "d6":  assertEquals(res6 , e.getValue());  break;
+				case "d7":  assertEquals(res7 , e.getValue());  break;
+				case "d8":  assertEquals(res8 , e.getValue());  break;
+				case "d9":  assertEquals(res9 , e.getValue());  break;
+				case "d10": assertEquals(res10, e.getValue()); break;
+				case "d11": assertEquals(res11, e.getValue()); break;
 				default:
 					// Do nothing
 					break;
